feat: add end-to-end investment analysis cockpit

+++ backend/src/main/java/com/immoradar/backend/deal/dto/CreateDealRequest.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 /**
  * DTO Java 25 Record pour la création d'un deal avec validation déclarative Jakarta.
  */
@@ -14,25 +16,25 @@ public record CreateDealRequest(
     String title,
 
     @Positive(message = "Le prix d'achat doit être strictement supérieur à zéro")
-    double price,
+    BigDecimal price,
 
     @Positive(message = "Le loyer mensuel estimé doit être strictement supérieur à zéro")
-    double monthlyRent,
+    BigDecimal monthlyRent,
 
     @PositiveOrZero(message = "Les charges mensuelles ne peuvent pas être négatives")
-    double monthlyCharges,
+    BigDecimal monthlyCharges,
 
     @PositiveOrZero(message = "La taxe foncière annuelle ne peut pas être négative")
-    double propertyTax,
+    BigDecimal propertyTax,
 
     @PositiveOrZero(message = "Le coût des travaux ne peut pas être négatif")
-    double renovationCost,
+    BigDecimal renovationCost,
 
     @NotBlank(message = "La localisation (ville ou département) est obligatoire")
     String location,
 
     @Positive(message = "La surface habitable doit être strictement supérieure à zéro")
-    double surface,
+    BigDecimal surface,
 
     @NotNull(message = "Le type de bien (Studio, Appartement, Immeuble, Maison) est obligatoire")
     PropertyType propertyType,
